service_receipt/pyproject.toml
[tool.poetry]
name = "receipt-generation"
version = "0.1.0"
description = ""
authors = ["ElenaGlu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fields = "^5.0.0"
psycopg2-binary = "^2.9.9"
pytest-django = "^4.8.0"
pytest = "^8.2.2"
celery = "^5.4.0"
pdfkit = "^1.0.0"
redis = "^5.0.6"
django = "4.2.13"
djangorestframework = "^3.15.2"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]

DJANGO_SETTINGS_MODULE = "service_receipt.settings"